+++ src/design/plone/contenttypes/tests/test_ct_persona.py
@@ -43,6 +43,7 @@ class TestPersona(unittest.TestCase):
                 "plone.basic",
                 "plone.locking",
                 "design.plone.contenttypes.behavior.additional_help_infos",
+                "design.plone.contenttypes.behavior.contatti_persona",
                 "plone.textindexer",
                 "plone.translatable",
                 "kitconcept.seo",
@@ -52,7 +53,7 @@ class TestPersona(unittest.TestCase):
 
     def test_persona_ct_title(self):
         portal_types = api.portal.get_tool(name="portal_types")
-        self.assertEqual("Persona", portal_types["Persona"].title)
+        self.assertEqual("Persona pubblica", portal_types["Persona"].title)
 
 
 class TestPersonaEndpoint(unittest.TestCase):