+++ src/design/plone/contenttypes/upgrades/upgrades.py
@@ -6,6 +6,7 @@ from copy import deepcopy
 from design.plone.contenttypes.controlpanels.settings import IDesignPloneSettings
 from design.plone.contenttypes.setuphandlers import remove_blocks_behavior
 from design.plone.contenttypes.upgrades.draftjs_converter import to_draftjs
+from design.plone.contenttypes.utils import create_default_blocks
 from plone import api
 from plone.app.textfield.value import RichTextValue
 from plone.app.upgrade.utils import installOrReinstallProduct
@@ -514,9 +515,6 @@ def to_volto13(context):  # noqa: C901
         for block in blocks.values():
             if block.get("@type", "") == "listing":
                 if block.get("template", False) and not block.get("variation", False):
-                    # import pdb
-
-                    # pdb.set_trace()
                     logger.error("- {}".format(url))
                 if block.get("template", False) and block.get("variation", False):
                     logger.error("- {}".format(url))
@@ -1193,9 +1191,6 @@ def migrate_pdc_and_incarico(context):
                 **kwargs,
             )
             intids = getUtility(IIntIds)
-            # import pdb
-
-            # pdb.set_trace()
             item.contact_info = [RelationValue(intids.getId(new_pdc))]
             fixed_total += 1
             commit()
@@ -1543,7 +1538,7 @@ def update_taxonomies(context):
         )
         for brain in brains:
             obj = brain.getObject()
-            obj_language = getattr(obj, "language", "it")
+            obj_language = getattr(obj, "language", "it") or "it"
             for taxonomy in TYPE_TO_TAXONOMIES_MAPPING[portal_type]:
                 old_value = getattr(obj, taxonomy)
                 if (
@@ -1778,3 +1773,36 @@ def update_patrocinato_da(self):
         )
         obj.reindexObject()
     logger.info(f"{colors.DARKCYAN} End of update {colors.ENDC}")
+
+
+def update_folder_for_gallery(self):
+    logger.info(f"{colors.DARKCYAN} Update events {colors.ENDC}")
+    pc = api.portal.get_tool(name="portal_catalog")
+    for brain in pc(portal_type="Event"):
+        evento = brain.getObject()
+
+        logger.info(f"{colors.DARKCYAN} Event: {evento.absolute_url()} {colors.ENDC}")
+        if "multimedia" in evento.keys():
+            renamed_event = api.content.rename(evento["multimedia"], new_id="immagini")
+            renamed_event.title = "Immagini"
+            renamed_event.reindexObject(idxs=["id", "title"])
+            logger.info(f"{colors.GREEN} Rename multimedia {colors.ENDC}")
+
+        if "video" not in evento.keys():
+            galleria_video = api.content.create(
+                container=evento,
+                type="Document",
+                title="Video",
+                id="video",
+            )
+            create_default_blocks(context=galleria_video)
+
+            # select  constraints
+            constraintsGalleriaVideo = ISelectableConstrainTypes(galleria_video)
+            constraintsGalleriaVideo.setConstrainTypesMode(1)
+            constraintsGalleriaVideo.setLocallyAllowedTypes(("Link",))
+
+            with api.env.adopt_roles(["Reviewer"]):
+                api.content.transition(obj=galleria_video, transition="publish")
+
+            logger.info(f"{colors.GREEN} Create video {colors.ENDC}")